requirements.txt
python-telegram-bot==20.7
google-cloud-dialogflow-cx==1.36.0
python-dotenv==1.0.0
